+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Other blocks are:
 ## Unreleased
 
 ### Added
+- **API** - Support of *hessvec*, *hessval*, *nvib* and second derivatives of energies in Gaussian log files.
+- **API** - New `get_hess_data` method in `DataFile` to get the eigenvectors (L) to transform from mass-weighted Cartesian to normal coordinates and frequencies.  This method acts as a wrapper and tries different ways to build the data.
 - **API** - Added qlabel *FCData:RedDim* and support in `parser.gaussian.glog` to extract information on normal-modes numbering in reduced-dimensionality schemes.
 
 ### Fixed